tests/wpt/tests/intersection-observer/scroll-margin-not-contained.html

<!DOCTYPE html>
<meta name="viewport" content="width=device-width,initial-scale=1">
<link rel="help" href="https://www.w3.org/TR/intersection-observer/#dom-intersectionobserver-scrollmargin">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./resources/intersection-observer-test-utils.js"></script>

<style>
#scroller { width: 100px; height: 100px; overflow: hidden; background-color: gray; position: absolute; }
#target { width: 50px; height: 50px; background-color: green; }
</style>

<div id=root>
  <div id=scroller>
    <div id=target></div>
  </div>
</div>

<script>
let entries = [];

window.onload = function() {
  runTestCycle(testIntersection, "Test scroll margin intersection");

  const observer = new IntersectionObserver(
    es => entries = entries.concat(es),
    {
      root: document.querySelector("#root"),
      scrollMargin: "10px"
    }
  );
  observer.observe(target);
};

function testIntersection() {
  assert_equals(entries.length, 1, "IntersectionObserverEntryCount");
  // Not insecting because root is not in the containing block chain of target.
  assert_false(entries[0].isIntersecting, "isIntersecting");
}
</script>
